% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average.R
\name{get_average}
\alias{get_average}
\title{get_average}
\usage{
get_average(data, type)
}
\arguments{
\item{data}{standard data format input}

\item{type}{type of function to average data, eg. median, mean...}
}
\value{
averaged groups
}
\description{
Average the data based on groups provided within data.frame
}
