% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.R
\name{get_fit}
\alias{get_fit}
\title{Fit grdR on growth curves}
\usage{
get_fit(
  inputData,
  groupingVariables,
  timeTreatment = 0,
  upperLimit = 0.9,
  upperLimitThreshold = 0.8,
  smoothData = T,
  orderConc = T,
  saveModel = F
)
}
\arguments{
\item{inputData}{input data.frame in standard format}

\item{groupingVariables}{variable to group (eg. controls, drugs, cells, etc..)}

\item{timeTreatment}{time of intervention treatment}

\item{upperLimit}{upperl in DRC: numeric upper limit parameters in the model}

\item{upperLimitThreshold}{threshold to evaluate GRD effect after model fit}

\item{smoothData}{should function to smooth data be applied?}

\item{orderConc}{should concentration be ordered based on effect?}

\item{saveModel}{should model parameters be returned?}
}
\description{
Fit grdR on growth curves
}
