% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logictic_fit.R
\name{get_logistic_fit}
\alias{get_logistic_fit}
\title{Fit model on data}
\usage{
get_logistic_fit(
  inputData,
  groupingVariables,
  upperLimit,
  lowerLimit = 1e-10,
  saveModel
)
}
\arguments{
\item{saveModel}{}
}
\description{
Fit model on data
}
