% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Get grdR output}
\usage{
get_output(outputData, metric)
}
\arguments{
\item{outputData}{output data from get_fit}

\item{metric}{return grdR or parameters}
}
\description{
Get grdR output
}
