% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get parameters from model}
\usage{
get_params(
  inputData,
  timeTreatment,
  upperLimit,
  upperLimitThreshold,
  orderConc,
  saveModel
)
}
\arguments{
\item{inputData}{input data in grdR format}

\item{timeTreatment}{time of drug treatment}

\item{upperLimit}{upper limit fitting function threshold}

\item{upperLimitThreshold}{upper limit fitting function threshold}

\item{orderConc}{order effect by concentrations}

\item{saveModel}{should model be saved}
}
\value{
fitted params and optionaly model
}
\description{
Get parameters from model
}
