% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smooth.R
\name{make_smooth}
\alias{make_smooth}
\title{Make data smooth}
\usage{
make_smooth(inputData, groupingVariables)
}
\arguments{
\item{inputData}{input data.frame in standard format}

\item{groupingVariables}{variable to group (eg. controls, drugs, cells, etc..)}
}
\value{
denoised data
}
\description{
Denoise data with smoothing function from mgcv::gam
}
