% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy.R
\name{make_tidy}
\alias{make_tidy}
\title{make data tidy}
\usage{
make_tidy(inputData, type)
}
\arguments{
\item{inputData}{input data.frame in standard format}

\item{type}{type of function to average data, eg. median, mean...}
}
\description{
@description  Make standard input data into tidy format/long format
}
