% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sensitivity.R
\name{model_growth}
\alias{model_growth}
\title{Create synthetic GRD data}
\usage{
model_growth(cell_tz, t, t_onset, k, maxeff, halfeff, conc, hill)
}
\arguments{
\item{cell_tz}{parameter}

\item{t}{parameter}

\item{t_onset}{parameter}

\item{k}{parameter}

\item{maxeff}{parameter}

\item{halfeff}{parameter}

\item{conc}{parameter}

\item{hill}{parameter}
}
\description{
Create synthetic GRD data
}
