% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot to check GRD on growth curves}
\usage{
plot_growth(inputData, groupingVariables, GRD, addLegend = FALSE)
}
\arguments{
\item{inputData}{Input data in GRD format for a single pertuabtion}

\item{groupingVariables}{variables to group the perturbation}

\item{GRD}{fitted GRD to be plotted as vertical line}

\item{addLegend}{shoudl concentration legend be added to plot}
}
\value{
a plot object
}
\description{
Plot to check GRD on growth curves
}
\examples{
inputData <- subset(sample_data,perturbation == "C")

outputGRD <- get_fit(inputData = inputData,groupingVariables = "perturbation",
    smoothData = FALSE, upperLimitThreshold = 1,
    timeTreatment = 1,upperLimit = 1, orderConc = TRUE,saveModel = FALSE)

plot_growth(inputData = inputData,groupingVariables = "perturbation", GRD = outputGRD$GRD)
}
