% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{grdR input sample data}
\format{
\subsection{\code{grd}}{

A data frame with 1,085 rows and 6 columns:
\describe{
\item{agent}{Country name}
\item{half eff}{parameter half eff}
\item{k}{parameter k}
\item{t_onset}{time of drug onset}
\item{concentration}{perturbation concentration}
\item{time}{time point}
\item{cell_count}{cell count}

}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
sample_data
}
\description{
Synthetic growth curve in the input format of the grdR package.
}
\keyword{datasets}
