% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logictic_fit.R
\name{get_logistic_fit}
\alias{get_logistic_fit}
\title{Fit model on data}
\usage{
get_logistic_fit(
  inputData,
  groupingVariables,
  upperLimit,
  lowerLimit = 1e-10,
  saveModel
)
}
\arguments{
\item{inputData}{input data}

\item{groupingVariables}{variables to group}

\item{upperLimit}{up boundary for fitting function}

\item{lowerLimit}{lower boundary for fitting function}

\item{saveModel}{should model be saved}
}
\value{
fitted logistic params/model
}
\description{
Fit model on data
}
