% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Get grdR output}
\usage{
get_output(outputData, metric)
}
\arguments{
\item{outputData}{output data from get_fit}

\item{metric}{return grdR or parameters}
}
\value{
output type from grd model
}
\description{
Get grdR output
}
\examples{
grd_output <- get_fit(sample_data, groupingVariables = "agent",
 smoothData = FALSE, upperLimitThreshold = 1,
 timeTreatment = 1,upperLimit = 1,orderConc = TRUE,saveModel = FALSE)

get_output(outputData = grd_output, metric = "grdR")

get_output(outputData = grd_output, metric = "parameters")
}
