% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get parameters from model}
\usage{
get_params(
  inputData,
  timeTreatment,
  upperLimit,
  upperLimitThreshold,
  orderConc,
  saveModel
)
}
\arguments{
\item{saveModel}{}
}
\description{
Get parameters from model
}
